introduce TemplateTrait
