+++ core-bundle/contao/templates/forms/form_captcha.html5
@@ -18,6 +18,8 @@
   <input type="hidden" name="<?= $this->name ?>_hash" value="<?= $this->hasErrors() ? $this->getHash() : '' ?>">
 
   <?php if (!$this->hasErrors()): ?>
+    <?php $this->addCspHash('style-src', 'display:none'); ?>
+    <?php $this->addCspSource('style-src', 'unsafe-hashes'); ?>
     <div style="display:none">
       <label for="ctrl_<?= $this->id ?>_hp">Do not fill in this field</label>
       <input type="text" name="<?= $this->name ?>_name" id="ctrl_<?= $this->id ?>_hp" value="">
